separate directories for configs and calada
